added version (auto/listed) to Client creation
When using docker-py 1.8.0 (released on Apr 6 2016), this script
can generate the error message about client API version being newer
than server (1.22 / 1.21). The fix for this, according to
[here](https://github.com/docker/docker-py/issues/1032) is to
specify the version or do specify 'auto' to use the server
version. I've added `--version` as a keyword argument before the
positional arguments for Client.